feat(45-02): Task 2 — field type fixes, validation, doc fixes
- DiffTruncation.original_bytes/truncated_bytes: usize → u64
- WorktreeStatus.ahead/behind: usize → u32
- stale_threshold_secs: validate non-zero at config load time
- Add test: validate_rejects_zero_stale_threshold_secs
- Add test: stale_threshold_secs_alias_backward_compat
- gate_runs field: document ID format (timestamp-6hex)
- EvaluatorCriterionResult.name: fix overstated doc
- fmt: fix formatting in spec/validate.rs
- Update diff-truncation and worktree-status schema snapshots

+++ crates/assay-types/src/gate_run.rs
@@ -68,9 +68,9 @@ inventory::submit! {
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, JsonSchema)]
 pub struct DiffTruncation {
     /// Byte size of the original diff before truncation.
-    pub original_bytes: usize,
+    pub original_bytes: u64,
     /// Byte size of the diff after truncation.
-    pub truncated_bytes: usize,
+    pub truncated_bytes: u64,
     /// Files included in the truncated diff.
     pub included_files: Vec<String>,
     /// Files omitted from the truncated diff (present in original, absent after truncation).
